+++ lib/BlackHole.js
@@ -1,14 +1,14 @@
 "use strict";
 
-const Writable = require('stream').Writable;
+import Writable from 'stream';
 
 class BlackHole extends Writable {
     constructor(options) {
         super(options);
     }
-    _write(chunk, encoding, callback) {
-        callback(); // eat the input
-    }
+    write() {} // eat the input
+    end() {}
 }
 
-module.exports = BlackHole;
+export { BlackHole };
+export default BlackHole;
