+++ lib/LinkCheckResult.js
@@ -11,4 +11,5 @@ class LinkCheckResult {
     }
 }
 
-module.exports = LinkCheckResult;
+export { LinkCheckResult };
+export default LinkCheckResult;
