+++ lib/proto/hash.js
@@ -1,10 +1,11 @@
 "use strict";
 
-const LinkCheckResult = require('../LinkCheckResult');
+import LinkCheckResult from '../LinkCheckResult.js';
 
-module.exports = {
-    check: function (link, opts, callback) {
-        const anchors = opts.anchors || [];
-        callback(null, new LinkCheckResult(opts, link, anchors.includes(link) ? 200 : 404, null));
-    }
-};
+function check(link, opts, callback) {
+    const anchors = opts.anchors || [];
+    callback(null, new LinkCheckResult(opts, link, anchors.includes(link) ? 200 : 404, null));
+}
+
+export { check };
+export default check;
