+++ lib/proto/https.js
@@ -1,3 +1,5 @@
 "use strict";
 
-module.exports = require('./http');
+import check from './http.mjs';
+export { check };
+export default check;
