+++ lib/proto/mailto.js
@@ -1,18 +1,29 @@
 "use strict";
 
-const Isemail = require('isemail');
-const LinkCheckResult = require('../LinkCheckResult');
+import emailValidator from 'node-email-verifier';
+import LinkCheckResult from '../LinkCheckResult.js';
 
-module.exports = {
-    check: function (link, opts, callback) {
-        const address = link
-                            .substr(7)      // strip "mailto:"
-                            .split('?')[0]; // trim ?subject=blah hfields
+async function check(link, opts, callback) {
+    const address = link
+                        .substr(7)      // strip "mailto:"
+                        .split('?')[0]; // trim ?subject=blah hfields
 
-        /* per RFC6068, the '?' is a reserved delimiter and email addresses containing '?' must be encoded,
-         * so it's safe to split on '?' and pick [0].
-         */
+    /* per RFC6068, the '?' is a reserved delimiter and email addresses containing '?' must be encoded,
+     * so it's safe to split on '?' and pick [0].
+     */
 
-        callback(null, new LinkCheckResult(opts, link, Isemail.validate(address) ? 200 : 400, null));
+    try {
+        if (!(await emailValidator(address, { checkMx: true, timeout: opts.timeout || '10s' }))) {
+            return callback(null, new LinkCheckResult(opts, link, 400, null));
+        }
+        return callback(null, new LinkCheckResult(opts, link, 200, null));
+    } catch (error) {
+        if (error.message.match(/timed out/)) {
+            return callback(null, new LinkCheckResult(opts, link, 0, { message: 'Domain MX lookup timed out', code: 'ECONNRESET' }));
+        }
+        return callback(null, new LinkCheckResult(opts, link, 0, error));
     }
-};
+}
+
+export { check };
+export default check;
