+++ package.json
@@ -3,6 +3,7 @@
   "version": "5.3.0",
   "description": "checks whether a hyperlink is alive (200 OK) or dead",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "pretest": "jshint index.js lib test",
     "test": "mocha -R spec --exit"
@@ -29,9 +30,9 @@
   "homepage": "https://github.com/tcort/link-check#readme",
   "dependencies": {
     "is-relative-url": "^4.0.0",
-    "isemail": "^3.2.0",
     "ms": "^2.1.3",
     "needle": "^3.3.1",
+    "node-email-verifier": "^2.0.0",
     "proxy-agent": "^6.4.0"
   },
   "devDependencies": {
